@@ -5,28 +5,67 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack serve --open",
-    "build": "webpack"
+    "dev": "cross-env TS_NODE_PROJECT=\"tsconfig.webpack.json\" NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=development webpack serve --open --config webpack-config/webpack.dev.ts",
+    "build": "cross-env TS_NODE_PROJECT=\"tsconfig.webpack.json\" NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=production webpack --config webpack-config/webpack.prod.ts",
+    "serve": "npm run build && http-server ./build"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@babel/core": "^7.19.3",
+    "@babel/plugin-transform-runtime": "^7.19.1",
+    "@babel/preset-env": "^7.19.3",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@types/glob": "^8.0.0",
+    "@types/node": "^18.7.23",
+    "@types/progress-bar-webpack-plugin": "^2.1.2",
+    "@types/speed-measure-webpack-plugin": "^1.3.4",
+    "@types/webpack": "^5.28.0",
+    "@types/webpack-bundle-analyzer": "^4.4.2",
+    "babel-loader": "^8.2.5",
+    "chalk": "^5.0.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "cross-env": "^7.0.3",
     "css-loader": "^6.2.0",
+    "glob": "^8.0.3",
     "html-webpack-plugin": "^5.3.2",
+    "http-server": "^14.1.1",
+    "mini-css-extract-plugin": "^2.6.1",
+    "pnp-webpack-plugin": "^1.7.0",
+    "progress-bar-webpack-plugin": "^2.1.0",
+    "purgecss-webpack-plugin": "^5.0.0",
+    "react-refresh": "^0.14.0",
     "sass-loader": "^12.1.0",
+    "scheduler": "^0.23.0",
+    "speed-measure-webpack-plugin": "^1.5.0",
     "style-loader": "^3.2.1",
+    "terser-webpack-plugin": "^5.3.6",
     "ts-loader": "^9.2.4",
-    "typescript": "^4.3.5",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.1.0",
+    "tslib": "^2.4.0",
+    "typescript": "^4.8.4",
     "webpack": "^5.46.0",
+    "webpack-bundle-analyzer": "^4.6.1",
     "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sass": "^1.36.0"
+  },
+  "ts-node": {
+    "esm": true,
+    "compilerOptions": {
+      "module": "CommonJS",
+      "moduleResolution": "NodeNext"
+    }
   }
 }
