@@ -1,17 +1,26 @@
 import React from 'react'
-import ReactDOM from 'react-dom' 
-import { Hello } from './hello.js';
+import { createRoot } from 'react-dom/client'
+import { Hello } from './hello'
 
-function Index () {
-    console.log('index')
-    return (
-        <div className='index'>
-            <p>hello index</p>
-            <Hello />
-        </div>
-    )
+function Index() {
+  console.log('index')
+  return (
+    <div className="index">
+      <p>hello index</p>
+      <Hello />
+    </div>
+  )
 }
 
-const element = <Index />
-const container = document.getElementById('container')
-ReactDOM.render(element, container);
+let container: HTMLElement
+
+if (document.getElementById('container')) {
+  container = document.getElementById('container') as HTMLElement
+} else {
+  container = document.createElement('div')
+  container.id = 'container'
+  document.body.appendChild(container)
+}
+
+const root = createRoot(container)
+root.render(<Index />)
